create facets.py
