+++ openlibrary/templates/search/work_search_selected_facets.html.jinja
@@ -1,7 +1,4 @@
-{# "Selected facets" chips shown above the search results.
-   Rendered by render_selected_search_facets() in openlibrary/plugins/openlibrary/partials.py,
-   which builds `other_chips` (see build_other_chips() there), the remove-facet URLs
-   (`del_facet_url`) and the search page's document title. #}
+{# "Selected facets" chips shown above the search results. #}
 {% set facet_tooltips = {
     'first_publish_year': _('First published in'),
     'publisher_facet': _('Published by'),
